requirements.txt
PyGithub>=2.1.0
PyYAML>=6.0
click>=8.1.0
requests>=2.28.0
gitpython>=3.1.0
